+++ great_expectations/data_context/store/__init__.py
@@ -1,4 +1,4 @@
-from great_expectations.util import verify_dynamic_loading_support
+# from great_expectations.util import verify_dynamic_loading_support
 
 from .store import Store  # isort:skip
 from .store_backend import (  # isort:skip
@@ -25,6 +25,6 @@ from .expectations_store import ExpectationsStore  # isort:skip
 from .validations_store import ValidationsStore  # isort:skip
 from .query_store import SqlAlchemyQueryStore  # isort:skip
 from .html_site_store import HtmlSiteStore  # isort:skip
-from .profiler_store import ProfilerStore  # isort:skip
 from .datasource_store import DatasourceStore  # isort:skip
 from .data_context_variables_store import DataContextVariablesStore  # isort:skip
+from .profiler_store import ProfilerStore  # isort:skip
