+++ great_expectations/rule_based_profiler/rule_based_profiler.py
@@ -22,7 +22,8 @@ from great_expectations.core.usage_statistics.usage_statistics import (
     usage_statistics_enabled_method,
 )
 from great_expectations.core.util import convert_to_json_serializable, nested_update
-from great_expectations.data_context.store import ProfilerStore
+
+# from great_expectations.data_context.store import ProfilerStore
 from great_expectations.data_context.types.resource_identifiers import (
     ConfigurationIdentifier,
     GeCloudIdentifier,
@@ -1029,7 +1030,7 @@ class BaseRuleBasedProfiler(ConfigPeer):
     @staticmethod
     def run_profiler(
         data_context: "BaseDataContext",  # noqa: F821
-        profiler_store: ProfilerStore,
+        profiler_store: "ProfilerStore",
         batch_list: Optional[List[Batch]] = None,
         batch_request: Optional[Union[BatchRequestBase, dict]] = None,
         name: Optional[str] = None,
@@ -1056,7 +1057,7 @@ class BaseRuleBasedProfiler(ConfigPeer):
     @staticmethod
     def run_profiler_on_data(
         data_context: "BaseDataContext",  # noqa: F821
-        profiler_store: ProfilerStore,
+        profiler_store: "ProfilerStore",
         batch_list: Optional[List[Batch]] = None,
         batch_request: Optional[Union[BatchRequestBase, dict]] = None,
         name: Optional[str] = None,
@@ -1087,7 +1088,7 @@ class BaseRuleBasedProfiler(ConfigPeer):
     def add_profiler(
         config: RuleBasedProfilerConfig,
         data_context: "BaseDataContext",  # noqa: F821
-        profiler_store: ProfilerStore,
+        profiler_store: "ProfilerStore",
         ge_cloud_id: Optional[str] = None,
     ) -> "RuleBasedProfiler":  # noqa: F821
         if not RuleBasedProfiler._check_validity_of_batch_requests_in_config(
@@ -1150,7 +1151,7 @@ class BaseRuleBasedProfiler(ConfigPeer):
     @staticmethod
     def get_profiler(
         data_context: "BaseDataContext",  # noqa: F821
-        profiler_store: ProfilerStore,
+        profiler_store: "ProfilerStore",
         name: Optional[str] = None,
         ge_cloud_id: Optional[str] = None,
     ) -> "RuleBasedProfiler":  # noqa: F821
@@ -1198,7 +1199,7 @@ class BaseRuleBasedProfiler(ConfigPeer):
 
     @staticmethod
     def delete_profiler(
-        profiler_store: ProfilerStore,
+        profiler_store: "ProfilerStore",
         name: Optional[str] = None,
         ge_cloud_id: Optional[str] = None,
     ) -> None:
@@ -1226,7 +1227,7 @@ class BaseRuleBasedProfiler(ConfigPeer):
 
     @staticmethod
     def list_profilers(
-        profiler_store: ProfilerStore,
+        profiler_store: "ProfilerStore",
         ge_cloud_mode: bool,
     ) -> List[str]:
         if ge_cloud_mode:
